chore: precompute rfc-index.txt API and refactoring of rfc extractor
