+++ rfc-html-extractor/src/tasks/rfc-index-txt.ts
@@ -1,0 +1,5 @@
+import type { RfcCommon } from "../../../client/app/utilities/rfc-validators.ts";
+
+export const generateRfcIndex = async (allRfcs: RfcCommon[]) => {
+    
+}
